requirements.txt
rdkit-pypi
streamlit
streamlit-ketcher
epam.indigo
python-docx
